requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.4
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
idna==2.10
kombu==5.1.0
Pillow==8.2.0
prompt-toolkit==3.0.19
pycparser==2.20
PyJWT==2.1.0
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5

# erlang-OPT==23.3
# rabbitmq-server==3.8.19